refactor(core/app): modularize app implementation
